+++ lib/src/nodes/changes/node_update.dart
@@ -20,4 +20,18 @@ class NodeUpdate extends NodeChange {
     required super.newState,
     super.oldState,
   });
+
+  @override
+  bool operator ==(Object other) {
+    if (other is! NodeUpdate) return false;
+    return newState == other.newState && oldState == other.oldState;
+  }
+
+  @override
+  int get hashCode => Object.hashAllUnordered(
+        <Object?>[
+          newState,
+          oldState,
+        ],
+      );
 }